+++ CHANGELOG.md
@@ -3,6 +3,7 @@ MumbleDJ Changelog
 
 ### January 25, 2015 -- `v2.3.0`
 * Added !currentsong command, which displays information about the song currently playing.
+* MumbleDJ now removes disconnected users from skiplists for playlists and songs within the SongQueue.
 
 ### January 19, 2015 -- `v2.2.11`
 * Fixed not being able to use the move command with channels with spaces in their name.